requirements.txt
Django==3.2.7
Pillow==9.4.0
gunicorn==20.1.0
whitenoise==5.3.0
sentry-sdk==1.3.1
requests==2.28.2
django-redis==5.0.0
pytz==2021.1
celery==5.1.2
celery-redbeat==2.0.0
psycopg2-binary==2.9.5
arrow==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-environ==0.6.0
django-storages==1.11.1
django-cors-headers==3.8.0
django-braces==1.14.0
django-extensions==3.1.3
django-post-office==3.5.3
django-crispy-forms==1.12.0
django-allauth==0.45.0
boto3==1.18.37
boto3-stubs==1.18.37
django-import-export==2.5.0
honeybadger==0.7.0
django-ses==2.2.1
djangoql==0.15.3
flake8==3.9.2
django-ckeditor==6.1.0
django-taggit==1.5.1
drf-yasg==1.20.0
django-filter==2.4.0
datapunt_keycloak_oidc==0.5.1